- security patch CAN-2003-0028 bug: The MIT Kerberos 5 implementation includes an RPC library derived from SUNRPC.  We have been notified that the xdrmem_getbytes() function contains faulty length checks.  These length checks are vulnerable to an integer overflow, which may be exploitable to create denials of service or to gain unauthorized access to sensitive information.
An attacker who has successfully authenticated to the Kerberos
administration daemon (kadmind) may be able to crash kadmind or induce
it to leak sensitive information, such as secret keys.  For the attack
to succeed, it is believed that the configuration of the kadmind
installation must allow it to successfully allocate more than INT_MAX
bytes of memory.

Changed files:
    krb5-security.patch -> 1.1